requirements.txt
django~=3.1.0
pillow
django-imagekit
pilkit
djangorestframework~=3.11.0
django-cors-headers
djangorestframework-jwt
django-filter

django-storages[azure]
gunicorn
psycopg2-binary